# backend/openapi.yaml
...
openapi: 3.0.3
info:
  title: Trackly API
  version: 1.0.0
  description: |
    OpenAPI specification for the Trackly backend.
    Interactive Swagger UI is available at `GET /docs`.
servers:
- url: http://localhost:4000
  description: Local development
security:
- bearerAuth: []
...
paths:
...
  "/api/recognition/audio":
    post:
      tags:
      - Recognition
      security: []
      summary: Recognize song from uploaded audio
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - audio
              properties:
                audio:
                  type: string
                  format: binary
      responses:
        '200':
          description: Recognition result
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/SongMetadata"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '404':
          description: No verified result found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '429':
          description: Recognition rate limit exceeded
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '500':
          "$ref": "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          enum:
          - INTERNAL_SERVER_ERROR
          - UNAUTHORIZED
          - FORBIDDEN
          - NOT_FOUND
          - INVALID_PAYLOAD
          - INVALID_USERNAME
          - INVALID_EMAIL
          - WEAK_PASSWORD
          - USERNAME_TAKEN
          - EMAIL_TAKEN
          - INVALID_CREDENTIALS
          - INVALID_PASSWORD
          - INVALID_NAME
          - MISSING_SONG_INFO
          - METHOD_REQUIRED
          - NO_VERIFIED_RESULT
          - PROVIDER_CONFIG_ERROR
          - AUDIO_FILE_REQUIRED
          - IMAGE_FILE_REQUIRED
          - AUDIO_RECOGNITION_FAILED
          - IMAGE_RECOGNITION_FAILED
          - TOO_MANY_RECOGNITION_REQUESTS
          - TOO_MANY_LOGIN_ATTEMPTS
          - SYNC_FAILED
          - GET_LIBRARY_FAILED
          - CREATE_FAILED
          - GET_FAILED
          - UPDATE_FAILED
          - ADD_SONG_FAILED
          - REMOVE_SONG_FAILED
          - DELETE_FAILED
        message:
          type: string
        details:
          nullable: true
        requestId:
          type: string
...
    SongMetadata:
      type: object
      required:
      - songName
      - artist
      - album
      - genre
      - releaseYear
      - confidenceScore
      - source
      - verificationStatus
      - platformLinks
      properties:
        songName:
          type: string
        artist:
          type: string
        album:
          type: string
        genre:
          type: string
        releaseYear:
          type: integer
          nullable: true
        confidenceScore:
          type: number
        youtubeVideoId:
          type: string
          nullable: true
        platformLinks:
          type: object
          additionalProperties:
            type: string
        source:
          type: string
          enum:
          - provider
          - ocr_fallback
        verificationStatus:
          type: string
          enum:
          - verified
          - not_found
...